+++ src/components/tina-markdown/embedded-elements/api-reference/utils.ts
@@ -1,3 +1,5 @@
+import { Endpoint } from "./types";
+
 export const resolveReference = (ref: string, definitions: any): any => {
   if (!ref || typeof ref !== "string" || !ref.startsWith("#/")) {
     return null;
@@ -68,4 +70,60 @@ export const generateExample = (
       }
       return null;
   }
+};
+
+export const extractEndpoints = (schemaJson: any) => {
+  const endpoints: Endpoint[] = [];
+  if (schemaJson.paths) {
+    for (const path of Object.keys(schemaJson.paths)) {
+      const pathObj = schemaJson.paths[path];
+      for (const method of Object.keys(pathObj)) {
+        if (method === "parameters") continue; // Skip path-level parameters
+
+        const operation = pathObj[method];
+
+        // Handle request body for OpenAPI 3.0 or Swagger 2.0
+        let requestBody: any = undefined;
+        if (operation.requestBody) {
+          requestBody = operation.requestBody;
+        } else if (operation.parameters?.some((p: any) => p.in === "body")) {
+          requestBody = {
+            content: {
+              "application/json": {
+                schema:
+                  operation.parameters.find((p: any) => p.in === "body")
+                    ?.schema || {},
+              },
+            },
+          };
+        }
+
+        // Filter out body parameters if we have a request body
+        const parameters = [
+          ...(pathObj.parameters || []), // Include path-level parameters
+          ...(operation.parameters || []),
+        ].filter((p) => {
+          // If we have a request body from a body parameter, filter out that parameter
+          if (requestBody && p.in === "body") {
+            return false;
+          }
+          return true;
+        });
+
+        endpoints.push({
+          path,
+          method: method.toUpperCase(),
+          summary: operation.summary || `${method.toUpperCase()} ${path}`,
+          description: operation.description,
+          operationId: operation.operationId,
+          parameters,
+          responses: operation.responses,
+          requestBody,
+          tags: operation.tags,
+          security: operation.security,
+        });
+      }
+    }
+  }
+  return endpoints;
 };
